+++ ThirdPartyAdapters/adcolony/adcolony/src/main/java/com/google/ads/mediation/adcolony/AdColonyMediationAdapter.java
@@ -1,6 +1,7 @@
 package com.google.ads.mediation.adcolony;
 
 import android.app.Activity;
+import android.app.Application;
 import android.content.Context;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -9,12 +10,14 @@ import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
 import com.adcolony.sdk.AdColony;
 import com.adcolony.sdk.AdColonyAppOptions;
-import com.adcolony.sdk.AdColonyCustomMessage;
-import com.adcolony.sdk.AdColonyCustomMessageListener;
+import com.adcolony.sdk.AdColonySignalsListener;
 import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.mediation.Adapter;
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback;
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
+import com.google.android.gms.ads.mediation.MediationBannerAd;
+import com.google.android.gms.ads.mediation.MediationBannerAdCallback;
+import com.google.android.gms.ads.mediation.MediationBannerAdConfiguration;
 import com.google.android.gms.ads.mediation.MediationConfiguration;
 import com.google.android.gms.ads.mediation.MediationInterstitialAd;
 import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback;
@@ -32,18 +35,16 @@ import com.jirbo.adcolony.BuildConfig;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 public class AdColonyMediationAdapter extends RtbAdapter {
 
   public static final String TAG = AdColonyMediationAdapter.class.getSimpleName();
   private static AdColonyAppOptions appOptions = new AdColonyAppOptions();
-  private static HashMap<String, String> bidResponseDetailsHashMap = new HashMap<>();
+
+  // Keeps a strong reference to the banner ad renderer, which loads ads asynchronously.
+  private AdColonyBannerRenderer adColonyBannerRenderer;
 
   // Keeps a strong reference to the interstitial ad renderer, which loads ads asynchronously.
   private AdColonyInterstitialRenderer adColonyInterstitialRenderer;
@@ -53,10 +54,10 @@ public class AdColonyMediationAdapter extends RtbAdapter {
 
   // region Error codes
   // AdColony adapter error domain.
-  public static final String ERROR_DOMAIN = "com.google.ads.mediation.adcolony";
+  private static final String ERROR_DOMAIN = "com.google.ads.mediation.adcolony";
 
   // AdColony SDK error domain.
-  public static final String ADCOLONY_SDK_ERROR_DOMAIN = "com.jirbo.adcolony";
+  private static final String ADCOLONY_SDK_ERROR_DOMAIN = "com.jirbo.adcolony";
 
   @IntDef(value = {
       ERROR_ADCOLONY_SDK,
@@ -67,10 +68,8 @@ public class AdColonyMediationAdapter extends RtbAdapter {
       ERROR_PRESENTATION_AD_NOT_LOADED,
       ERROR_CONTEXT_NOT_ACTIVITY,
   })
-
   @Retention(RetentionPolicy.SOURCE)
   public @interface AdapterError {
-
   }
 
   /**
@@ -108,19 +107,19 @@ public class AdColonyMediationAdapter extends RtbAdapter {
    */
   public static final int ERROR_CONTEXT_NOT_ACTIVITY = 106;
 
-  // TODO: Migrate all users to return an AdError instead.
-  @Deprecated
   @NonNull
-  public static String createAdapterError(@AdapterError int error, @NonNull String errorMessage) {
-    return String.format(Locale.US, "%d: %s", error, errorMessage);
+  public static AdError createAdapterError(@AdapterError int error, @NonNull String errorMessage) {
+    return new AdError(error, errorMessage, ERROR_DOMAIN);
   }
 
-  // TODO: Migrate all users to return an AdError instead.
-  @Deprecated
   @NonNull
-  public static String createSdkError() {
-    return String.format(Locale.US, "%d: %s", ERROR_ADCOLONY_SDK,
-        "AdColony SDK returned a failure callback");
+  public static AdError createSdkError() {
+    return createSdkError(ERROR_ADCOLONY_SDK, "AdColony SDK returned a failure callback.");
+  }
+
+  @NonNull
+  public static AdError createSdkError(@AdapterError int error, @NonNull String errorMessage) {
+    return new AdError(error, errorMessage, ADCOLONY_SDK_ERROR_DOMAIN);
   }
   // endregion
 
@@ -128,6 +127,7 @@ public class AdColonyMediationAdapter extends RtbAdapter {
    * {@link Adapter} implementation
    */
   @Override
+  @NonNull
   public VersionInfo getVersionInfo() {
     String versionString = BuildConfig.ADAPTER_VERSION;
     String[] splits = versionString.split("\\.");
@@ -147,6 +147,7 @@ public class AdColonyMediationAdapter extends RtbAdapter {
   }
 
   @Override
+  @NonNull
   public VersionInfo getSDKVersionInfo() {
     String sdkVersion = AdColony.getSDKVersion();
     String[] splits = sdkVersion.split("\\.");
@@ -165,16 +166,17 @@ public class AdColonyMediationAdapter extends RtbAdapter {
   }
 
   @Override
-  public void initialize(Context context,
-      final InitializationCompleteCallback initializationCompleteCallback,
-      List<MediationConfiguration> mediationConfigurations) {
-
-    if (!(context instanceof Activity)) {
-      initializationCompleteCallback.onInitializationFailed("AdColony SDK requires an " +
-          "Activity context to initialize");
+  public void initialize(
+          @NonNull Context context,
+          @NonNull final InitializationCompleteCallback initializationCompleteCallback,
+          @NonNull List<MediationConfiguration> mediationConfigurations
+  ) {
+    if (!(context instanceof Activity) && !(context instanceof Application)) {
+      AdError error = createAdapterError(ERROR_CONTEXT_NOT_ACTIVITY,
+              "AdColony SDK requires an Activity or Application context to initialize.");
+      initializationCompleteCallback.onInitializationFailed(error.toString());
       return;
     }
-    Activity activity = (Activity) context;
 
     HashSet<String> appIDs = new HashSet<>();
     ArrayList<String> zoneList = new ArrayList<>();
@@ -198,21 +200,26 @@ public class AdColonyMediationAdapter extends RtbAdapter {
     String appID;
     int count = appIDs.size();
     if (count <= 0) {
-      initializationCompleteCallback.onInitializationFailed("Missing or invalid AdColony app ID.");
+      AdError error = createAdapterError(ERROR_INVALID_SERVER_PARAMETERS,
+              "Missing or invalid AdColony app ID.");
+      initializationCompleteCallback.onInitializationFailed(error.toString());
       return;
     }
 
     appID = appIDs.iterator().next();
     if (count > 1) {
-      String logMessage = String.format("Multiple '%s' entries found: %s. " +
-              "Using '%s' to initialize the AdColony SDK.",
-          AdColonyAdapterUtils.KEY_APP_ID, appIDs.toString(), appID);
+      String logMessage = String.format(
+              "Multiple '%s' entries found: %s. Using '%s' to initialize the AdColony SDK.",
+              AdColonyAdapterUtils.KEY_APP_ID,
+              appIDs.toString(),
+              appID
+      );
       Log.w(TAG, logMessage);
     }
 
     // Always set mediation network info.
     appOptions.setMediationNetwork(AdColonyAppOptions.ADMOB, BuildConfig.ADAPTER_VERSION);
-    AdColonyManager.getInstance().configureAdColony(activity, appOptions, appID,
+    AdColonyManager.getInstance().configureAdColony(context, appOptions, appID,
         zoneList, new InitializationListener() {
           @Override
           public void onInitializeSuccess() {
@@ -222,29 +229,43 @@ public class AdColonyMediationAdapter extends RtbAdapter {
           @Override
           public void onInitializeFailed(@NonNull AdError error) {
             // TODO: Forward the AdError object once available.
-            initializationCompleteCallback.onInitializationFailed(error.getMessage());
+            initializationCompleteCallback.onInitializationFailed(error.toString());
           }
         });
   }
 
   @Override
-  public void loadInterstitialAd(
-      MediationInterstitialAdConfiguration interstitialAdConfiguration,
-      MediationAdLoadCallback<MediationInterstitialAd,
-          MediationInterstitialAdCallback> mediationAdLoadCallback) {
-    String requestedZone = interstitialAdConfiguration.getServerParameters()
-        .getString(AdColonyAdapterUtils.KEY_ZONE_ID);
-    adColonyInterstitialRenderer = new AdColonyInterstitialRenderer(requestedZone);
-    adColonyInterstitialRenderer.requestInterstitial(mediationAdLoadCallback);
+  public void loadRtbBannerAd(
+          @NonNull MediationBannerAdConfiguration bannerAdConfiguration,
+          @NonNull MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> mediationAdLoadCallback
+  ) {
+    adColonyBannerRenderer = new AdColonyBannerRenderer(bannerAdConfiguration, mediationAdLoadCallback);
+    adColonyBannerRenderer.render();
+  }
+
+  @Override
+  public void loadRtbInterstitialAd(
+          @NonNull MediationInterstitialAdConfiguration interstitialAdConfiguration,
+          @NonNull MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback> mediationAdLoadCallback
+  ) {
+    adColonyInterstitialRenderer = new AdColonyInterstitialRenderer(interstitialAdConfiguration, mediationAdLoadCallback);
+    adColonyInterstitialRenderer.render();
+  }
+
+  @Override
+  public void loadRtbRewardedAd(
+          @NonNull MediationRewardedAdConfiguration rewardedAdConfiguration,
+          @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> mediationAdLoadCallback
+  ) {
+    loadRewardedAd(rewardedAdConfiguration, mediationAdLoadCallback);
   }
 
   @Override
   public void loadRewardedAd(
-      MediationRewardedAdConfiguration rewardedAdConfiguration,
-      MediationAdLoadCallback<MediationRewardedAd,
-          MediationRewardedAdCallback> mediationAdLoadCallback) {
-    adColonyRewardedRenderer =
-        new AdColonyRewardedRenderer(rewardedAdConfiguration, mediationAdLoadCallback);
+          @NonNull MediationRewardedAdConfiguration rewardedAdConfiguration,
+          @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback> mediationAdLoadCallback
+  ) {
+    adColonyRewardedRenderer = new AdColonyRewardedRenderer(rewardedAdConfiguration, mediationAdLoadCallback);
     adColonyRewardedRenderer.render();
   }
 
@@ -253,26 +274,20 @@ public class AdColonyMediationAdapter extends RtbAdapter {
   }
 
   @Override
-  public void collectSignals(RtbSignalData rtbSignalData, SignalCallbacks signalCallbacks) {
-    String zone, signals = "";
-    AdColony.addCustomMessageListener(new AdColonyCustomMessageListener() {
+  public void collectSignals(@NonNull RtbSignalData rtbSignalData, @NonNull final SignalCallbacks signalCallbacks) {
+    AdColony.collectSignals(new AdColonySignalsListener() {
       @Override
-      public void onAdColonyCustomMessage(AdColonyCustomMessage adColonyCustomMessage) {
-        try {
-          JSONObject jsonObject = new JSONObject(adColonyCustomMessage.getMessage());
-          String zoneID = jsonObject.getString("zone");
-          bidResponseDetailsHashMap.put(zoneID, adColonyCustomMessage.getMessage());
-        } catch (JSONException e) {
-          e.printStackTrace();
-        }
+      public void onSuccess(String signals) {
+        signalCallbacks.onSuccess(signals);
       }
-    }, "bid");
-    if (bidResponseDetailsHashMap != null && bidResponseDetailsHashMap.size() > 0) {
-      Bundle serverParameters = rtbSignalData.getConfiguration().getServerParameters();
-      zone = serverParameters.getString(AdColonyAdapterUtils.KEY_ZONE_ID);
-      signals = bidResponseDetailsHashMap.get(zone);
-    }
-    signalCallbacks.onSuccess(signals);
-  }
 
+      @Override
+      public void onFailure() {
+        AdError error = createSdkError(ERROR_ADCOLONY_SDK,
+                "Failed to get signals from AdColony.");
+        Log.e(TAG, error.getMessage());
+        signalCallbacks.onFailure(error);
+      }
+    });
+  }
 }