+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/google/ads/mediation/vungle/VungleMediationAdapter.java
@@ -41,9 +41,9 @@ public class VungleMediationAdapter extends Adapter
   private AdConfig mAdConfig;
   private String mUserID;
   private String mPlacement;
-  private Handler mHandler = new Handler(Looper.getMainLooper());
+  private final Handler mHandler = new Handler(Looper.getMainLooper());
 
-  private static HashMap<String, WeakReference<VungleMediationAdapter>> mPlacementsInUse =
+  private static final HashMap<String, WeakReference<VungleMediationAdapter>> mPlacementsInUse =
       new HashMap<>();
 
   private MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
@@ -52,7 +52,7 @@ public class VungleMediationAdapter extends Adapter
 
   @Override
   public VersionInfo getVersionInfo() {
-    String versionString = BuildConfig.VERSION_NAME;
+    String versionString = BuildConfig.ADAPTER_VERSION;
     String[] splits = versionString.split("\\.");
 
     if (splits.length >= 4) {
@@ -225,7 +225,7 @@ public class VungleMediationAdapter extends Adapter
   }
 
   /**
-   * {@link LoadAdCallback} implemenatation from Vungle
+   * {@link LoadAdCallback} implemenatation from Vungle.
    */
   @Override
   public void onAdLoad(final String placementId) {
@@ -253,8 +253,6 @@ public class VungleMediationAdapter extends Adapter
           public void run() {
             if (mMediationRewardedAdCallback != null) {
               mMediationRewardedAdCallback.onAdOpened();
-              mMediationRewardedAdCallback.onVideoStart();
-              mMediationRewardedAdCallback.reportAdImpression();
             }
           }
         });
@@ -262,9 +260,7 @@ public class VungleMediationAdapter extends Adapter
 
   @Override
   @Deprecated
-  public void onAdEnd(
-      final String placementId,
-      final boolean wasSuccessfulView,
+  public void onAdEnd(final String placementId, final boolean wasSuccessfulView,
       final boolean wasCallToActionClicked) {
   }
 
@@ -323,7 +319,7 @@ public class VungleMediationAdapter extends Adapter
           @Override
           public void run() {
             if (mMediationAdLoadCallback != null) {
-              Log.w(TAG, "Failed to load ad from Vungle", throwable);
+              Log.w(TAG, "Failed to load ad from Vungle.", throwable);
               mMediationAdLoadCallback.onFailure(throwable.getLocalizedMessage());
             }
 
@@ -337,7 +333,8 @@ public class VungleMediationAdapter extends Adapter
 
   @Override
   public void onAdViewed(String placementId) {
-    // "no-op , to be mapped to respective adapter events in future release"
+    mMediationRewardedAdCallback.onVideoStart();
+    mMediationRewardedAdCallback.reportAdImpression();
   }
 
   /**