+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleInterstitialAdapter.java
@@ -24,8 +24,8 @@ import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
-import android.widget.RelativeLayout;
 import androidx.annotation.Keep;
+import com.google.ads.mediation.vungle.VungleBannerAd;
 import com.google.ads.mediation.vungle.VungleInitializer;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdSize;
@@ -36,6 +36,10 @@ import com.google.android.gms.ads.mediation.MediationBannerListener;
 import com.google.android.gms.ads.mediation.MediationInterstitialAdapter;
 import com.google.android.gms.ads.mediation.MediationInterstitialListener;
 import com.vungle.warren.AdConfig;
+import com.vungle.warren.LoadAdCallback;
+import com.vungle.warren.PlayAdCallback;
+import com.vungle.warren.Vungle;
+import com.vungle.warren.error.VungleException;
 import java.util.ArrayList;
 
 /**
@@ -53,23 +57,19 @@ public class VungleInterstitialAdapter
   private String mPlacementForPlay;
 
   // banner/MREC
-  private volatile RelativeLayout adLayout;
   private MediationBannerListener mMediationBannerListener;
-  private VungleBannerAdapter mBannerRequest;
+  private VungleBannerAdapter vungleBannerAdapter;
 
   @Override
-  public void requestInterstitialAd(
-      Context context,
-      MediationInterstitialListener mediationInterstitialListener,
-      Bundle serverParameters,
-      MediationAdRequest mediationAdRequest,
-      Bundle mediationExtras) {
+  public void requestInterstitialAd(Context context,
+      MediationInterstitialListener mediationInterstitialListener, Bundle serverParameters,
+      MediationAdRequest mediationAdRequest, Bundle mediationExtras) {
 
     AdapterParametersParser.Config config;
     try {
       config = AdapterParametersParser.parse(mediationExtras, serverParameters);
     } catch (IllegalArgumentException e) {
-      Log.w(TAG, "Failed to load ad from Vungle", e);
+      Log.w(TAG, "Failed to load ad from Vungle.", e);
       if (mediationInterstitialListener != null) {
         mediationInterstitialListener.onAdFailedToLoad(
             VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
@@ -82,7 +82,7 @@ public class VungleInterstitialAdapter
 
     mPlacementForPlay = mVungleManager.findPlacement(mediationExtras, serverParameters);
     if (TextUtils.isEmpty(mPlacementForPlay)) {
-      Log.w(TAG, "Failed to load ad from Vungle: Missing or Invalid Placement ID");
+      Log.w(TAG, "Failed to load ad from Vungle: Missing or Invalid Placement ID.");
       mMediationInterstitialListener.onAdFailedToLoad(
           VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
       return;
@@ -116,115 +116,123 @@ public class VungleInterstitialAdapter
       if (mMediationInterstitialListener != null) {
         mMediationInterstitialListener.onAdLoaded(VungleInterstitialAdapter.this);
       }
-    } else if (mVungleManager.isValidPlacement(mPlacementForPlay)) {
-      mVungleManager.loadAd(
-          mPlacementForPlay,
-          new VungleListener() {
-            @Override
-            void onAdAvailable() {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdLoaded(VungleInterstitialAdapter.this);
-              }
-            }
+      return;
+    }
 
-            @Override
-            void onAdFailedToLoad(int errorCode) {
-              Log.w(TAG, "Failed to load ad from Vungle: " + errorCode);
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdFailedToLoad(
-                    VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_NO_FILL);
-              }
-            }
-          });
-    } else { // passed Placement Id is not what Vungle's SDK gets back after init/config
+    // Placement ID is not what Vungle's SDK gets back after init/config.
+    if (!mVungleManager.isValidPlacement(mPlacementForPlay)) {
       if (mMediationInterstitialListener != null) {
         mMediationInterstitialListener.onAdFailedToLoad(
             VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
       }
+      return;
     }
+
+    Vungle.loadAd(mPlacementForPlay, new LoadAdCallback() {
+      @Override
+      public void onAdLoad(String placementID) {
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdLoaded(VungleInterstitialAdapter.this);
+        }
+      }
+
+      @Override
+      public void onError(String placementID, VungleException exception) {
+        Log.w(TAG, "Failed to load ad from Vungle: " + exception.getLocalizedMessage());
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdFailedToLoad(
+              VungleInterstitialAdapter.this, exception.getExceptionCode());
+        }
+      }
+    });
   }
 
   @Override
   public void showInterstitial() {
-    if (mVungleManager != null) {
-      mVungleManager.playAd(
-          mPlacementForPlay,
-          mAdConfig,
-          new VungleListener() {
-            @Override
-            void onAdClick(String placementId) {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdClicked(VungleInterstitialAdapter.this);
-              }
-            }
+    Vungle.playAd(mPlacementForPlay, mAdConfig, new PlayAdCallback() {
+      @Override
+      public void onAdStart(String placementID) {
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdOpened(VungleInterstitialAdapter.this);
+        }
+      }
 
-            @Override
-            void onAdEnd(String placementId) {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdClosed(VungleInterstitialAdapter.this);
-              }
-            }
+      @Override
+      public void onAdEnd(String placementID, boolean completed, boolean isCTAClicked) {
+        // Deprecated, no-op.
+      }
 
-            @Override
-            void onAdLeftApplication(String placementId) {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdLeftApplication(VungleInterstitialAdapter.this);
-              }
-            }
+      @Override
+      public void onAdEnd(String placementID) {
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdClosed(VungleInterstitialAdapter.this);
+        }
+      }
 
-            @Override
-            void onAdStart(String placement) {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdOpened(VungleInterstitialAdapter.this);
-              }
-            }
+      @Override
+      public void onAdClick(String placementID) {
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdClicked(VungleInterstitialAdapter.this);
+        }
+      }
 
-            @Override
-            void onAdFail(String placement) {
-              if (mMediationInterstitialListener != null) {
-                mMediationInterstitialListener.onAdClosed(VungleInterstitialAdapter.this);
-              }
-            }
-          });
-    }
+      @Override
+      public void onAdRewarded(String placementID) {
+        // No-op for interstitial ads.
+      }
+
+      @Override
+      public void onAdLeftApplication(String placementID) {
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdLeftApplication(VungleInterstitialAdapter.this);
+        }
+      }
+
+      @Override
+      public void onError(String placementID, VungleException exception) {
+        Log.w(TAG, "Failed to play ad from Vungle: " + exception.getLocalizedMessage());
+        if (mMediationInterstitialListener != null) {
+          mMediationInterstitialListener.onAdClosed(VungleInterstitialAdapter.this);
+        }
+      }
+
+      @Override
+      public void onAdViewed(String id) {
+        // No-op.
+      }
+    });
   }
 
   @Override
   public void onDestroy() {
     Log.d(TAG, "onDestroy: " + hashCode());
-    if (mBannerRequest != null) {
-      mBannerRequest.destroy(adLayout);
-      mBannerRequest = null;
+    if (vungleBannerAdapter != null) {
+      vungleBannerAdapter.destroy();
+      vungleBannerAdapter = null;
     }
-    adLayout = null;
   }
 
   // banner
   @Override
   public void onPause() {
     Log.d(TAG, "onPause");
-    if (mBannerRequest != null) {
-      mBannerRequest.updateVisibility(false);
+    if (vungleBannerAdapter != null) {
+      vungleBannerAdapter.updateVisibility(false);
     }
   }
 
   @Override
   public void onResume() {
     Log.d(TAG, "onResume");
-    if (mBannerRequest != null) {
-      mBannerRequest.updateVisibility(true);
+    if (vungleBannerAdapter != null) {
+      vungleBannerAdapter.updateVisibility(true);
     }
   }
 
   @Override
-  public void requestBannerAd(
-      Context context,
-      final MediationBannerListener mediationBannerListener,
-      Bundle serverParameters,
-      AdSize adSize,
-      MediationAdRequest mediationAdRequest,
-      Bundle mediationExtras) {
-    Log.d(TAG, "requestBannerAd");
+  public void requestBannerAd(Context context,
+      final MediationBannerListener mediationBannerListener, Bundle serverParameters, AdSize adSize,
+      MediationAdRequest mediationAdRequest, Bundle mediationExtras) {
     mMediationBannerListener = mediationBannerListener;
 
     AdapterParametersParser.Config config;
@@ -242,12 +250,8 @@ public class VungleInterstitialAdapter
     mVungleManager = VungleManager.getInstance();
 
     String placementForPlay = mVungleManager.findPlacement(mediationExtras, serverParameters);
-    Log.d(
-        TAG,
-        "requestBannerAd for Placement: "
-            + placementForPlay
-            + " ###  Adapter instance: "
-            + this.hashCode());
+    Log.d(TAG, "requestBannerAd for Placement: " + placementForPlay
+        + " ### Adapter instance: " + this.hashCode());
 
     if (TextUtils.isEmpty(placementForPlay)) {
       String message = "Failed to load ad from Vungle: Missing or Invalid Placement ID.";
@@ -266,109 +270,31 @@ public class VungleInterstitialAdapter
       return;
     }
 
-    // Create the adLayout wrapper with the requested ad size, as Vungle's ad uses MATCH_PARENT for
-    // its dimensions.
-    adLayout =
-        new RelativeLayout(context) {
-          @Override
-          protected void onAttachedToWindow() {
-            super.onAttachedToWindow();
-            if (mBannerRequest != null) {
-              mBannerRequest.attach();
-            }
-          }
-
-          @Override
-          protected void onDetachedFromWindow() {
-            super.onDetachedFromWindow();
-            if (mBannerRequest != null) {
-              mBannerRequest.detach();
-            }
-          }
-        };
-    int adLayoutHeight = adSize.getHeightInPixels(context);
-    // If the height is 0 (e.g. for inline adaptive banner requests), use the closest supported size
-    // as the height of the adLayout wrapper.
-    if (adLayoutHeight <= 0) {
-      float density = context.getResources().getDisplayMetrics().density;
-      adLayoutHeight = Math.round(adConfig.getAdSize().getHeight() * density);
-    }
-    RelativeLayout.LayoutParams adViewLayoutParams =
-        new RelativeLayout.LayoutParams(adSize.getWidthInPixels(context), adLayoutHeight);
-    adLayout.setLayoutParams(adViewLayoutParams);
-
-    mBannerRequest =
-        mVungleManager.getBannerRequest(placementForPlay, config.getRequestUniqueId(), adConfig);
-    if (mBannerRequest == null) {
-      // Adapter does not support multiple Banner instances playing for same placement except for
-      // Refresh
+    // Adapter does not support multiple Banner instances playing for same placement except for
+    // refresh.
+    String uniqueRequestId = config.getRequestUniqueId();
+    if (!mVungleManager.canRequestBannerAd(placementForPlay, uniqueRequestId)) {
       mMediationBannerListener.onAdFailedToLoad(
           VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
       return;
     }
 
-    mBannerRequest.setAdLayout(adLayout);
-    mBannerRequest.setVungleListener(mVungleBannerListener);
+    vungleBannerAdapter = new VungleBannerAdapter(placementForPlay, uniqueRequestId, adConfig,
+        VungleInterstitialAdapter.this);
+    Log.d(TAG, "New banner adapter: " + vungleBannerAdapter + "; size: " + adConfig.getAdSize());
+
+    VungleBannerAd vungleBanner = new VungleBannerAd(placementForPlay, vungleBannerAdapter);
+    mVungleManager.registerBannerAd(placementForPlay, vungleBanner);
 
     Log.d(TAG, "Requesting banner with ad size: " + adConfig.getAdSize());
-    mBannerRequest.requestBannerAd(context, config.getAppId());
+    vungleBannerAdapter
+        .requestBannerAd(context, config.getAppId(), adSize, mMediationBannerListener);
   }
-
-  private VungleListener mVungleBannerListener =
-      new VungleListener() {
-        @Override
-        void onAdClick(String placementId) {
-          if (mMediationBannerListener != null) {
-            mMediationBannerListener.onAdClicked(VungleInterstitialAdapter.this);
-            mMediationBannerListener.onAdOpened(VungleInterstitialAdapter.this);
-          }
-        }
-
-        @Override
-        void onAdEnd(String placementId) {
-          // No-op.
-        }
-
-        @Override
-        void onAdLeftApplication(String placementId) {
-          if (mMediationBannerListener != null) {
-            mMediationBannerListener.onAdLeftApplication(VungleInterstitialAdapter.this);
-          }
-        }
-
-        @Override
-        void onAdAvailable() {
-          if (mMediationBannerListener != null) {
-            mMediationBannerListener.onAdLoaded(VungleInterstitialAdapter.this);
-          }
-        }
-
-        @Override
-        void onAdStart(String placement) {
-          // let's load it again to mimic auto-cache, don't care about errors
-          if (mBannerRequest != null) {
-            mBannerRequest.preCache();
-          }
-        }
-
-        @Override
-        void onAdFail(String placement) {
-          Log.w(TAG, "Ad playback error Placement: " + placement + ";" + mBannerRequest);
-        }
-
-        @Override
-        void onAdFailedToLoad(int errorCode) {
-          Log.w(TAG, "Failed to load ad from Vungle: " + errorCode + ";" + mBannerRequest);
-          if (mMediationBannerListener != null) {
-            mMediationBannerListener.onAdFailedToLoad(VungleInterstitialAdapter.this, errorCode);
-          }
-        }
-      };
 
   @Override
   public View getBannerView() {
     Log.d(TAG, "getBannerView # instance: " + hashCode());
-    return adLayout;
+    return vungleBannerAdapter.getAdLayout();
   }
 
   private boolean hasBannerSizeAd(Context context, AdSize adSize, AdConfig adConfig) {