+++ ThirdPartyAdapters/adcolony/CHANGELOG.md
@@ -1,5 +1,30 @@
 ## AdColony Android Mediation Adapter Changelog
 
+#### Version 4.5.0.0
+- Added support for banners advanced bidding.
+- Verified compatibility with AdColony SDK version 4.5.0.
+- Updated the minimum required Google Mobile Ads SDK version to 20.0.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 20.0.0.
+- AdColony SDK version 4.5.0.
+
+#### Version 4.4.1.0
+- Verified compatibility with AdColony SDK version 4.4.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.7.0.
+- AdColony SDK version 4.4.1.
+
+#### Version 4.4.0.0
+- Verified compatibility with AdColony SDK version 4.4.0.
+- Added support of AdColony's `collectSignals()` method for open bidding.
+- Updated the minimum required Google Mobile Ads SDK version to 19.7.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.7.0.
+- AdColony SDK version 4.4.0.
+
 #### Version 4.3.0.0
 - Verified compatibility with AdColony SDK version 4.3.0.
 - Updated the minimum required Google Mobile Ads SDK version to 19.5.0.